mostly done
